refactor(audit): resolve all descoped follow-ups from PR 562 audit
- borders/lifecycle: gate markWindowSnapped overlay on current-desktop and
  wire desktopChanged -> updateAllBorders so borders rebuild reliably on switch
- geometry: thread IZoneLayoutRegistry through the daemon geometry helpers so
  drag-preview / empty-zone overlay / zone-detection geometry reflect the
  current-context rule gap override (matches commit geometry); add
  resolveContextGaps to IZoneLayoutRegistry (stub-safe default) and extract the
  shared contextGapOverrideMap converter
- phosphor-zones: split batch-assignment machinery into layoutregistry_batch.cpp
  (both TUs now <800 lines)
- effect: precompute hasOpacityRules() so the per-frame opacity resolve skips the
  WindowQuery build when no enabled SetOpacity rule exists
- workspaces: add ActivityManager::currentActivityOrEmpty(), collapse the 4
  duplicated (mgr && isAvailable()) activity guards
- windowrule: broaden hasHexColor to the standard QColor hex shapes
  (#RGB/#RRGGBB/#AARRGGBB) the consumer parses; sync docstring + test
