+++ kwin-effect/plasmazoneseffect/paint_pipeline.cpp
@@ -260,7 +260,7 @@ void PlasmaZonesEffect::prePaintWindow(KWin::RenderView* view, KWin::EffectWindo
         // the whole animation. setTranslucent() clears the opaque region
         // so every frame fully recomposites under the window.
         data.setTranslucent();
-    } else if (w && !m_shaderManager.animationRuleSet().isEmpty()) {
+    } else if (w && m_shaderManager.hasOpacityRules()) {
         // SetOpacity-only case: a rule may dim this window via
         // `data.setOpacity()` in paintWindow below, but with neither an
         // animation, shader transition, nor restore suppression in
@@ -381,7 +381,7 @@ void PlasmaZonesEffect::paintWindow(const KWin::RenderTarget& renderTarget, cons
     // to keep default-state cost to two pointer reads; without it, every
     // paintWindow call churns the rule evaluator's per-window cache for
     // zero benefit.
-    if (!m_shaderManager.animationRuleSet().isEmpty()) {
+    if (m_shaderManager.hasOpacityRules()) {
         const QString winClass = w->windowClass();
         if (!isOwnOverlayClass(winClass) && !isPlasmaShellSurface(winClass)) {
             // Reuse the per-frame cache populated by prePaintWindow above.
