+++ src/core/daemongeometryresolver.cpp
@@ -13,40 +13,15 @@ namespace PlasmaZones {
 
 QVariantMap DaemonGeometryResolver::contextGapOverrideFor(const QString& screenId) const
 {
-    namespace PSK = PhosphorEngine::PerScreenSnappingKey;
-    QVariantMap map;
     if (!m_layoutRegistry || screenId.isEmpty()) {
-        return map;
+        return {};
     }
     const int virtualDesktop = m_currentVirtualDesktop ? m_currentVirtualDesktop() : 0;
     const QString activity = m_currentActivity ? m_currentActivity() : QString();
-    const PhosphorZones::ContextGapOverride gaps =
-        m_layoutRegistry->resolveContextGaps(screenId, virtualDesktop, activity);
-    if (gaps.isEmpty()) {
-        return map;
-    }
-    if (gaps.zonePadding) {
-        map.insert(QString(PSK::ZonePadding), *gaps.zonePadding);
-    }
-    if (gaps.outerGap) {
-        map.insert(QString(PSK::OuterGap), *gaps.outerGap);
-    }
-    if (gaps.usePerSideOuterGap) {
-        map.insert(QString(PSK::UsePerSideOuterGap), *gaps.usePerSideOuterGap);
-    }
-    if (gaps.outerGapTop) {
-        map.insert(QString(PSK::OuterGapTop), *gaps.outerGapTop);
-    }
-    if (gaps.outerGapBottom) {
-        map.insert(QString(PSK::OuterGapBottom), *gaps.outerGapBottom);
-    }
-    if (gaps.outerGapLeft) {
-        map.insert(QString(PSK::OuterGapLeft), *gaps.outerGapLeft);
-    }
-    if (gaps.outerGapRight) {
-        map.insert(QString(PSK::OuterGapRight), *gaps.outerGapRight);
-    }
-    return map;
+    // Translation to the PerScreenSnappingKey-shaped map is shared with the
+    // preview/query geometry helpers via GeometryUtils::contextGapOverrideMap.
+    return GeometryUtils::contextGapOverrideMap(
+        m_layoutRegistry->resolveContextGaps(screenId, virtualDesktop, activity));
 }
 
 int DaemonGeometryResolver::resolveZonePadding(PhosphorZones::Layout* layout, const QString& screenId) const
